added suicidal_index in users table

+++ db/migrate/20160829093611_add_column_users.rb
@@ -1,5 +1,5 @@
 class AddColumnUsers < ActiveRecord::Migration[5.0]
   def change
-  	add_column :users, :suicidal_index, :integer
+  	add_column :users, :suicidal_index, :float
   end
 end
